requirements.txt
joblib==1.1.0
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
plotly==5.3.1
scikit_learn==1.0.1
streamlit==1.1.0

